Init user service (#39)
* Copy user service

https://github.com/CS3219-AY2425S1/PeerPrep-UserService

* Dockerise user-service

Got user service to work in docker. It can only interface with an atlas cloud server currently.

* Dockerise user-service

User-service in docker can only communicate with an atlas mongodb server.

* Merge with main

Some files should have been committed in the previous commit. They are
mostly to do with the dockerisation process. The merge with main might
have messed with it.

* Create .gitignore

Somehow .gitignore was not commited with the last commit

* Fix POST causing SEGSEGV

Apparently bcrypt library cannot run on Docker's architecture. Change
bcrypt to bcryptjs for compatibility.

* Add test

Test does not currently work

* Basic user service

Not connected to the login interface yet. Commiting to make a clean Pr.

* Move user service folder

* Fix minor user service issues

* Remove orphan history file
* Move `index.js` and `server.js` to the correct directory
* Reduce use of port

* Change login to use username

* Fix user env typo

---------

Co-authored-by: samuelim01 <samuelim01@gmail.com>