Connect frontend with user service (#47)
* Copy user service

https://github.com/CS3219-AY2425S1/PeerPrep-UserService

* Dockerise user-service

Got user service to work in docker. It can only interface with an atlas cloud server currently.

* Dockerise user-service

User-service in docker can only communicate with an atlas mongodb server.

* Merge with main

Some files should have been committed in the previous commit. They are
mostly to do with the dockerisation process. The merge with main might
have messed with it.

* Create .gitignore

Somehow .gitignore was not commited with the last commit

* Fix POST causing SEGSEGV

Apparently bcrypt library cannot run on Docker's architecture. Change
bcrypt to bcryptjs for compatibility.

* Add test

Test does not currently work

* Basic user service

Not connected to the login interface yet. Commiting to make a clean Pr.

* Move user service folder

* Fix minor user service issues

* Remove orphan history file
* Move `index.js` and `server.js` to the correct directory
* Reduce use of port

* Change login to use username

* Fix user env typo

* Enable login

Frontend is now able to POST to user login endpoint and receive a JWT
token and some user details. These are stored in localStorage.

* Same commit as before

Some new files were untracked.

* Implement login and register

Right now if register is successful, it will auto login.

Login if successful will bring users to the '/' directory.

If user is already logged in, attemptes to reach '/account' will be
immediately redirected to '/'

* Rename files

To keep it consistent with main, I have renamed files that are not part
of app from filename to "_filename" out of app. I also cleaned up the
outdated user-service because it was renamed already.

* Fix wrong merge

There was a duplicated of "user" in compose.yml.

* Init guards

Used ng generate guard to generate these.

* Implement auth-guard and interceptors

Removed those _guards and implemented as a service instead. Added a
guard to '/questions' so only logged in users with tokens that have not
expired can access it. If they are not logged in, users will be directed
to '/login' to log in.

* Fix prettier

Was magically missing from last commit.

* Fix silly issues

- Removed user network (errornous merging).
- Removed duplicate providers

* Fix linting

Added interface for user service response.

* Temp commit

Trying to figure out how to get status code from HttpClient with
HttpResponse.

* Fix error handling

Realised that the error interceptor was mishandling the errors. Now it
will throw an Error object with properties .message and .cause. Where
message is the original error.message, and cause is just the original
error itself.

* Fix linting

* Minor fixes

* Move interceptors into _interceptors
* Clean user and user service response models
* Fix auth service unhandled cases
* Minor phrasing fixes

* Clean JWT interceptor

* Use API_CONFIG instead of environment.ts

* Oops

* Tie in gateway changes properly

---------

Co-authored-by: samuelim01 <samuelim01@gmail.com>